refactor: use custom hook
